Refactored BuildURL helper to use URLSearchParams serializer;
