+++ README.md
@@ -332,7 +332,7 @@ These are the available config options for making requests. Only the `url` is re
   headers: {'X-Requested-With': 'XMLHttpRequest'},
 
   // `params` are the URL parameters to be sent with the request
-  // Must be a plain object or a URLSearchParams object
+  // Must be a plain object or a AxiosURLSearchParams object
   params: {
     ID: 12345
   },
@@ -346,7 +346,7 @@ These are the available config options for making requests. Only the `url` is re
   // `data` is the data to be sent as the request body
   // Only applicable for request methods 'PUT', 'POST', 'DELETE , and 'PATCH'
   // When no `transformRequest` is set, must be of one of the following types:
-  // - string, plain object, ArrayBuffer, ArrayBufferView, URLSearchParams
+  // - string, plain object, ArrayBuffer, ArrayBufferView, AxiosURLSearchParams
   // - Browser only: FormData, File, Blob
   // - Node only: Stream, Buffer
   data: {
