+++ lib/helpers/toFormData.js
@@ -55,6 +55,10 @@ var predicates = utils.toFlatObject(utils, {}, null, function filter(prop) {
  **/
 
 function toFormData(obj, formData, options) {
+  if (!utils.isObject(obj)) {
+    throw new TypeError('target must be an object');
+  }
+
   // eslint-disable-next-line no-param-reassign
   formData = formData || new (envFormData || FormData)();
